projects/map/src/lib/helpers/map-size.helper.ts
export class MapSizeHelper {

  public static getFormattedLength(size?: number): string {
    if (!size) {
      return '';
    }
    if (size > 100) {
      return (Math.round(size / 1000 * 100) / 100) + ' ' + 'km';
    } else {
      return (Math.round(size * 100) / 100) + ' ' + 'm';
    }
  }

  public static getFormattedArea(size?: number): string {
    if (!size) {
      return '';
    }
    if (size > 10000) {
      return (Math.round(size / 1000000 * 100) / 100) + ' ' + 'km';
    } else {
      return (Math.round(size * 100) / 100) + ' ' + 'm';
    }
  }

}
